Update J-Box Schematic.pdf
Changed Schematic to reflect pinout used on connector old pin numbers are on outside and new pin numbers are on inside. Pinout expected to be changed in the future so that pin 1 is pin one. But for now this takes some of the guess work out of wiring the J-Box